POC: proxy node for super node by rewriting JanusGraphVertexStep
In order to make the changes as little as possible, this POC almost only modified JanusGraphVertexStep,
which is the entrypoint to all vertex centric queries. This POC relies on the label of proxy node to judge
whether a transformation is needed, which is not efficient. A more efficient way is to use internal long id
to identify a proxy node. In order to avoid breaking changes, this commit does not do that.

Node that in this commit, canonical nodes and proxy nodes are connected implicitly, i.e. via property.